+++ xlnt/include/xlnt/utils/datetime.hpp
@@ -1,4 +1,4 @@
-// Copyright (c) 2014-2018 Thomas Fussell
+// Copyright (c) 2014-2017 Thomas Fussell
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -54,7 +54,7 @@ struct XLNT_API datetime
     /// a date and the fractional part into a time according to date::from_number
     /// and time::from_number.
     /// </summary>
-    static datetime from_number(double number, calendar base_date);
+    static datetime from_number(long double number, calendar base_date);
 
     /// <summary>
     /// Returns a datetime equivalent to the ISO-formatted string iso_string.
@@ -84,7 +84,7 @@ struct XLNT_API datetime
     /// <summary>
     /// Returns this datetime as a number of seconds since 1900 or 1904 (depending on base_date provided).
     /// </summary>
-    double to_number(calendar base_date) const;
+    long double to_number(calendar base_date) const;
 
     /// <summary>
     /// Returns true if this datetime is equivalent to comparand.
